+++ app/javascript/vue/tasks/observation_matrices/new/components/columns/copyDescriptors.vue
@@ -109,8 +109,7 @@ import { ActionNames } from '../../store/actions/actions'
 import { GetterNames } from '../../store/getters/getters'
 import { ObservationMatrix, ObservationMatrixColumnItem } from 'routes/endpoints'
 import {
-  GetMatrixObservationColumns,
-  CreateColumnItem
+  GetMatrixObservationColumns
 } from '../../request/resources'
 
 export default {
@@ -150,7 +149,7 @@ export default {
       handler (newVal) {
         if (newVal) {
           this.isLoading = true
-          ObservationMatrix.all().then(response => {
+          ObservationMatrix.where({ per: 500 }).then(response => {
             response.body.splice(response.body.findIndex(item => this.matrixId === item.id), 1)
             this.observationMatrices = response.body
             this.isLoading = false