+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_11_25_201511) do
+ActiveRecord::Schema.define(version: 2021_12_28_013444) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "fuzzystrmatch"
@@ -1085,7 +1085,7 @@ ActiveRecord::Schema.define(version: 2021_11_25_201511) do
     t.datetime "updated_at", null: false
     t.string "type"
     t.index ["created_by_id"], name: "labels_created_by_id_index"
-    t.index ["label_object_type", "label_object_id"], name: "index_labels_on_label_object_type_and_label_object_id"
+    t.index ["label_object_type", "label_object_id"], name: "index_labels_on_label_object"
     t.index ["project_id"], name: "index_labels_on_project_id"
     t.index ["updated_by_id"], name: "labels_updated_by_id_index"
   end
@@ -1349,8 +1349,8 @@ ActiveRecord::Schema.define(version: 2021_11_25_201511) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["created_by_id"], name: "index_origin_relationships_on_created_by_id"
-    t.index ["new_object_type", "new_object_id"], name: "index_origin_relationships_on_new_object_type_and_new_object_id"
-    t.index ["old_object_type", "old_object_id"], name: "index_origin_relationships_on_old_object_type_and_old_object_id"
+    t.index ["new_object_type", "new_object_id"], name: "index_origin_relationships_on_new_object"
+    t.index ["old_object_type", "old_object_id"], name: "index_origin_relationships_on_old_object"
     t.index ["project_id"], name: "index_origin_relationships_on_project_id"
     t.index ["updated_by_id"], name: "index_origin_relationships_on_updated_by_id"
   end
@@ -1417,7 +1417,9 @@ ActiveRecord::Schema.define(version: 2021_11_25_201511) do
     t.integer "year_died"
     t.integer "year_active_start"
     t.integer "year_active_end"
+    t.index ["cached"], name: "index_people_on_cached"
     t.index ["created_by_id"], name: "index_people_on_created_by_id"
+    t.index ["last_name"], name: "index_people_on_last_name"
     t.index ["type"], name: "index_people_on_type"
     t.index ["updated_by_id"], name: "index_people_on_updated_by_id"
   end
@@ -1436,7 +1438,7 @@ ActiveRecord::Schema.define(version: 2021_11_25_201511) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["created_by_id"], name: "index_pinboard_items_on_created_by_id"
-    t.index ["pinned_object_type", "pinned_object_id"], name: "index_pinboard_items_on_pinned_object_type_and_pinned_object_id"
+    t.index ["pinned_object_type", "pinned_object_id"], name: "index_pinboard_items_on_pinned_object"
     t.index ["position"], name: "index_pinboard_items_on_position"
     t.index ["project_id"], name: "index_pinboard_items_on_project_id"
     t.index ["updated_by_id"], name: "index_pinboard_items_on_updated_by_id"