+++ lib/export/dwca.rb
@@ -16,14 +16,15 @@ module Export
       '2021-10-15 17:00:00.000000 -0500',    # Minor  Excludes footprintWKT, and references to GeographicArea in gazetteer; new form of media links
       '2021-11-04 17:00:00.000000 -0500',    # Minor  Removes '|', fixes some mappings
       '2021-11-08 13:00:00.000000 -0500',    # PENDING: Minor  Adds depth mappings
-      '2021-11-30 13:00:00.000000 -0500'     # Fix inverted long,lat 
+      '2021-11-30 13:00:00.000000 -0500',    # Fix inverted long,lat 
+      '2022-01-21 16:30:00.000000 -0500'     # basisOfRecord can now be FossilSpecimen; occurrenceId exporting; adds redundant time fields
     ]
 
     # @param record_scope [ActiveRecord::Relation]
     #   a relation that returns DwcOccurrence records
     # @return [Download]
     #   the download object containing the archive
-    def self.download_async(record_scope, request = nil, predicate_extension_params = nil)
+    def self.download_async(record_scope, request = nil, predicate_extension_params: {})
       name = "dwc-a_#{DateTime.now}.zip"
 
       download = ::Download::DwcArchive.create!(