+++ spec/lib/export/dwca/data_spec.rb
@@ -27,30 +27,28 @@ describe Export::Dwca::Data, type: :model, group: :darwin_core do
         end
       end
 
-      after do
-        data.cleanup
-      end
+      after { data.cleanup }
 
       let(:csv) { CSV.parse(data.csv, headers: true, col_sep: "\t") }
-      let(:headers) { ['basisOfRecord', 'individualCount' ] }
+      let(:headers) { [ 'basisOfRecord', 'individualCount', 'occurrenceID' ] } # id, and non-standard DwC colums are handled elsewhere
 
       context 'various scopes' do
         specify 'with .where clauses' do
           s = scope.where('id > 1')
           d = Export::Dwca::Data.new(core_scope: s)
-          expect(d.csv_headers).to contain_exactly(*headers)
+          expect(d.meta_fields).to contain_exactly(*headers)
         end
 
         specify 'with .order clauses' do
           s = scope.order(:basisOfRecord)
           d = Export::Dwca::Data.new(core_scope: s)
-          expect(d.csv_headers).to contain_exactly(*headers)
+          expect(d.meta_fields).to contain_exactly(*headers)
         end
 
         specify 'with .join clauses' do
           s = scope.collection_objects_join
           d = Export::Dwca::Data.new(core_scope: s)
-          expect(d.csv_headers).to contain_exactly(*headers)
+          expect(d.meta_fields).to contain_exactly(*headers)
         end
       end
 
@@ -63,11 +61,11 @@ describe Export::Dwca::Data, type: :model, group: :darwin_core do
       end
 
       specify 'generated headers are restricted to data' do
-        expect(csv.headers).to contain_exactly('id', 'basisOfRecord', 'individualCount')
+        expect(csv.headers).to contain_exactly('id', 'occurrenceID', 'basisOfRecord', 'individualCount')
       end
 
-      specify '#csv_headers can be returned, and exclude id' do
-        expect(data.csv_headers).to contain_exactly(*headers)
+      specify '#meta_fields can be returned, and exclude id' do
+        expect(data.meta_fields).to contain_exactly(*headers)
       end
 
       context 'files' do